atualizações da pasta e da lista 01

+++ listas/unidade-01/lista-01/atv04.c
@@ -12,19 +12,19 @@ int main()
     scanf("%d", &num2);
     
     res = num1 + num2;
-    printf("A soma de %d + %d é de: %.1f\n", num1, num2, res);
+    printf("A soma de %d + %d é de: %.0f\n", num1, num2, res);
     
     res = num1 - num2;
-    printf("A diferença de %d - %d é de: %.1f\n", num1, num2, res);
+    printf("A diferença de %d - %d é de: %.0f\n", num1, num2, res);
     
     res = num1 * num2;
-    printf("O produto de %d * %d é de: %.1f\n", num1, num2, res);
+    printf("O produto de %d * %d é de: %.0f\n", num1, num2, res);
     
-    res = num1 / num2;
+    res = (float)num1 / num2;
     printf("A divisão de %d / %d é de: %.1f\n", num1, num2, res);
     
     res = num1 % num2;
-    printf("O resto de %d por %d é de: %.1f\n", num1, num2, res);
+    printf("O resto de %d por %d é de: %.0f\n", num1, num2, res);
     
    return 0; 
 }
